Added a slider cell. Thanks to Jason Booth.
The Continuous option allows you to specify if the slider should wait until the user lets go to call it's onChanged selector, or call it every time the slider moves. The slider will size itself like other controls in the table view unless you provide the FixedSize key; setting this value will force the slider to be a specific size; useful if you have many sliders in a control and want them to line up. The onValueChanged key is the selector you want to call on your delegate when the value is changed, e.g.:

- (void) sliderChanged:(id)sender

where sender is the slider.